Implement user authorization

+++ routes/events.py
@@ -7,14 +7,15 @@ from fastapi import (
     status
 )
 from sqlmodel import select
+from auths.authenticate import authenticate
 from models.db_connect import get_session
-from schemas.events import Event, EventUpdate
+from schemas.events import Event, EventRes, EventUpdate
 
 event_router: APIRouter = APIRouter(prefix="/events", tags=["Events"])
 
 
-@event_router.get("/", response_model=List[Event])
-async def get_events(session=Depends(get_session)) -> List[Event]:
+@event_router.get("/", response_model=List[EventRes])
+async def get_events(session=Depends(get_session)) -> List[EventRes]:
     """Return all the available events."""
     statement = select(Event)
     events = session.exec(statement).all()
@@ -22,7 +23,10 @@ async def get_events(session=Depends(get_session)) -> List[Event]:
 
 
 @event_router.get("/{event_id}", response_model=Event)
-async def get_an_event(event_id: int, session=Depends(get_session)) -> Event:
+async def get_an_event(
+    event_id: int, user: str = Depends(authenticate),
+    session=Depends(get_session)
+) -> Event:
     """Return an event by id."""
     event = session.get(Event, event_id)
     if event:
@@ -34,8 +38,12 @@ async def get_an_event(event_id: int, session=Depends(get_session)) -> Event:
 
 
 @event_router.post("/create", status_code=status.HTTP_201_CREATED)
-async def create_event(event: Event, session=Depends(get_session)) -> dict:
+async def create_event(
+    event: Event, user: str = Depends(authenticate),
+    session=Depends(get_session)
+) -> dict:
     """Create a new event."""
+    event.creator = user["user"]
     session.add(event)
     session.commit()
     session.refresh(event)
@@ -46,13 +54,23 @@ async def create_event(event: Event, session=Depends(get_session)) -> dict:
     "/delete/{event_id}",
     # status_code=status.HTTP_204_NO_CONTENT
 )
-async def delete_an_event(event_id: int, session=Depends(get_session)):
+async def delete_an_event(
+    event_id: int, user: str = Depends(authenticate),
+    session=Depends(get_session)
+):
     """Delete an event."""
     event = session.get(Event, event_id)
-    if event:
+    if event and user["user"] == event.creator:
         session.delete(event)
         session.commit()
         return {"message": "Event deleted successfully"}
+
+    if event and user["user"] != event.creator:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Permission denied!"
+        )
+
     raise HTTPException(
         status_code=status.HTTP_404_NOT_FOUND,
         detail="Event not found"
@@ -66,11 +84,13 @@ async def delete_an_event(event_id: int, session=Depends(get_session)):
 async def update_event(
     event_id: int,
     new_data: EventUpdate,
+    user: str = Depends(authenticate),
     session=Depends(get_session)
 ) -> dict:
     """Update an events."""
     event = session.get(Event, event_id)
-    if event:
+
+    if event and user["user"] == event.creator:
         event_data = new_data.dict(exclude_unset=True)
         for k, v in event_data.items():
             setattr(event, k, v)
@@ -78,6 +98,13 @@ async def update_event(
         session.commit()
         session.refresh(event)
         return event
+
+    if event and user["user"] != event.creator:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Permission denied!"
+        )
+
     raise HTTPException(
         status_code=status.HTTP_404_NOT_FOUND,
         detail="Event not found"
